Interview getting saved/deleted/updated now.

+++ src/hooks/useApplicationData.js
@@ -13,22 +13,22 @@ const useApplicationData = () => {
 
   // Change number of spots when a new interview is booked
   // or an existing interview is deleted
-  const updateSpots = () => {
-    setState((prev) => {
-      const { day, appointments, days } = { ...prev };
-      let spots = 0;
-      for (const d of days) {
-        if (d.name === day) {
-          for (const id of d.appointments) {
-            appointments[id].interview === null && spots++;
-            d.spots = spots;
-          }
-        }
-      }
-      return { ...prev, days };
-    });
-    // console.log("State after update",state);
-  };
+  // const updateSpots = () => {
+  //   setState((prev) => {
+  //     const { day, appointments, days } = { ...prev };
+  //     let spots = 0;
+  //     for (const d of days) {
+  //       if (d.name === day) {
+  //         for (const id of d.appointments) {
+  //           appointments[id].interview === null && spots++;
+  //           d.spots = spots;
+  //         }
+  //       }
+  //     }
+  //     return { ...prev, days };
+  //   });
+  //   // console.log("State after update",state);
+  // };
 
   // Book an interview and update state to reflect change
   const bookInterview = (id, interview) => {
@@ -39,7 +39,7 @@ const useApplicationData = () => {
     appointments.forEach((appt) => {
       if (appt._id === id) appt = appointment;
     });
-    console.log("appointments", appointments);
+    // console.log("appointments", appointments);
 
     // const appointment = {
     //   ...state.appointments[id],
@@ -65,23 +65,22 @@ const useApplicationData = () => {
 
   // Cancel an interview and update state to reflect change
   const cancelInterview = (id) => {
-    const appointment = {
-      ...state.appointments[id],
-      interview: null,
-    };
+    console.log("I got a request to cancel this appointment (useAppData):", id);
 
-    const appointments = {
-      ...state.appointments,
-      [id]: appointment,
-    };
+    const appointments = [...state.appointments];
+    const appointment = appointments.find((appt) => appt._id === id);
+    if (!appointment) return null;
+    console.log("I found that appointment:", appointment);
+ 
 
     const deleteAppointment = axios.delete(`/api/appointments/${id}`);
 
     return deleteAppointment.then((response) => {
+      console.log("This response was received from the database", response);
       response.status === 200 && setState({ ...state, appointments });
       // console.log("Appointments after book interview", appointments)
       // console.log(response);
-      updateSpots();
+      // updateSpots();
     });
   };
